chore: Inject modules in the constructor.
